CDEC-438: Split sin binned Reporting between core and infra
Anything that was required by the blockchain package had to go to
`core` while everything else went to `infra`.
