+++ core/src/Pos/Core/Reporting.hs
@@ -1,0 +1,7 @@
+module Pos.Core.Reporting
+       ( module X
+       ) where
+
+import           Pos.Core.Reporting.MemState as X
+import           Pos.Core.Reporting.Methods as X
+import           Pos.Core.Reporting.Metrics as X
