+++ delegation/cardano-sl-delegation.cabal
@@ -40,7 +40,6 @@ library
                      , cardano-sl-db
                      , cardano-sl-lrc
                      , cardano-sl-networking
-                     , cardano-sl-sinbin
                      , cardano-sl-util
                      , conduit
                      , ether
