+++ infra/cardano-sl-infra.cabal
@@ -29,9 +29,7 @@ library
                         Pos.Infra.Reporting.Exceptions
                         Pos.Infra.Reporting.Health.Types
                         Pos.Infra.Reporting.Http
-                        Pos.Infra.Reporting.MemState
                         Pos.Infra.Reporting.Methods
-                        Pos.Infra.Reporting.Metrics
                         Pos.Infra.Reporting.NodeInfo
                         Pos.Infra.Reporting.Statsd
                         Pos.Infra.Reporting.Wlog
@@ -127,11 +125,13 @@ library
                       , cardano-sl-binary
                       , cardano-sl-core
                       , cardano-sl-crypto
+                      , cardano-sl-db
                       , cardano-sl-networking
                       , cardano-sl-ssc
-                      , cardano-sl-sinbin
                       , cardano-sl-util
                       , cardano-report-server >= 0.4.5
+                      , clock
+                      , conduit
                       , containers
                       , directory
                       , dns
@@ -140,8 +140,11 @@ library
                       , ekg-wai
                       , ether >= 0.5
                       , exceptions
+                      , filepath
                       , formatting
                       , hashable
+                      , http-client
+                      , http-client-tls
                       , iproute
                       , kademlia
                       , lens
@@ -150,11 +153,13 @@ library
                       , network-info
                       , network-transport
                       , network-transport-tcp
+                      , lzma-conduit
                       , optparse-applicative
                       , safe-exceptions
                       , serokell-util >= 0.1.3.4
                       , stm
-                      , clock
+                      , tar
+                      , time
                       , tagged
                       , text
                       , time-units
