+++ infra/src/Pos/Infra/Diffusion/Types.hs
@@ -25,7 +25,7 @@ import           Pos.Core.Update (UpId, UpdateProposal, UpdateVote)
 import           Pos.Infra.Communication.Types.Protocol (NodeId)
 import           Pos.Infra.Diffusion.Subscription.Status (SubscriptionStates,
                      emptySubscriptionStates)
-import           Pos.Sinbin.Reporting (HealthStatus (..))
+import           Pos.Infra.Reporting.Health.Types (HealthStatus (..))
 
 
 data DiffusionHealth = DiffusionHealth {
