--- infra/src/Pos/Infra/Reporting/MemState.hs
@@ -1,5 +1,0 @@
-module Pos.Infra.Reporting.MemState
-       ( module X
-       ) where
-
-import           Pos.Sinbin.Reporting.MemState as X
