--- infra/src/Pos/Infra/Reporting/Metrics.hs
@@ -1,5 +1,0 @@
-module Pos.Infra.Reporting.Metrics
-       ( module X
-       ) where
-
-import           Pos.Sinbin.Reporting.Metrics as X
