+++ infra/src/Pos/Infra/Slotting/Util.hs
@@ -19,7 +19,7 @@ module Pos.Infra.Slotting.Util
          -- * Waiting for system start
        , waitSystemStart
 
-         -- * Re-exported from the Sinbin
+         -- * Re-exported from the Infra
        , ActionTerminationPolicy (..)
        , OnNewSlotParams (..)
        , defaultOnNewSlotParams
@@ -43,11 +43,11 @@ import           Pos.Core.Slotting (ActionTerminationPolicy (..),
                      getCurrentSlotFlat, getEpochSlottingDataM,
                      getSystemStartM, lookupEpochSlottingData)
 import           Pos.Infra.Recovery.Info (MonadRecoveryInfo, recoveryInProgress)
+import           Pos.Infra.Reporting (MonadReporting, reportOrLogE)
 import           Pos.Infra.Shutdown (HasShutdownContext)
 import           Pos.Infra.Slotting.Class (MonadSlots (..))
 import           Pos.Infra.Slotting.Error (SlottingError (..))
 import           Pos.Infra.Slotting.Impl.Util (slotFromTimestamp)
-import           Pos.Sinbin.Reporting (MonadReporting, reportOrLogE)
 import           Pos.Util.Util (maybeThrow)
