+++ lib/src/Pos/Network/Block/Retrieval.hs
@@ -25,13 +25,13 @@ import           Pos.Core (Block, HasHeaderHash (..), HeaderHash, difficultyL,
                      isMoreDifficult)
 import           Pos.Core.Block (BlockHeader)
 import           Pos.Core.Chrono (NE, OldestFirst (..), _OldestFirst)
+import           Pos.Core.Reporting (HasMisbehaviorMetrics)
 import           Pos.Crypto (ProtocolMagic, shortHashF)
 import qualified Pos.DB.BlockIndex as DB
 import           Pos.Infra.Communication.Protocol (NodeId)
 import           Pos.Infra.Diffusion.Types (Diffusion)
 import qualified Pos.Infra.Diffusion.Types as Diffusion
-import           Pos.Infra.Reporting (HasMisbehaviorMetrics, reportOrLogE,
-                     reportOrLogW)
+import           Pos.Infra.Reporting (reportOrLogE, reportOrLogW)
 import           Pos.Network.Block.Logic (BlockNetLogicException (..),
                      handleBlocks, triggerRecovery)
 import           Pos.Network.Block.RetrievalQueue (BlockRetrievalQueueTag,
