+++ lib/src/Pos/WorkMode/Class.hs
@@ -27,6 +27,7 @@ import           Pos.Context (BlockRetrievalQueue, BlockRetrievalQueueTag,
                      HasSscContext, StartTime, TxpGlobalSettings)
 import           Pos.Core (HasConfiguration, HasPrimaryKey)
 import           Pos.Core.JsonLog (CanJsonLog)
+import           Pos.Core.Reporting (HasMisbehaviorMetrics, MonadReporting)
 import           Pos.DB.Class (MonadDB, MonadGState)
 import           Pos.DB.Rocks (MonadRealDB)
 import           Pos.Delegation.Class (MonadDelegation)
@@ -34,7 +35,6 @@ import           Pos.Delegation.Configuration (HasDlgConfiguration)
 import           Pos.Infra.DHT.Real.Param (KademliaParams)
 import           Pos.Infra.Network.Types (HasNodeType, NetworkConfig)
 import           Pos.Infra.Recovery.Info (MonadRecoveryInfo)
-import           Pos.Infra.Reporting (HasMisbehaviorMetrics, MonadReporting)
 import           Pos.Infra.Shutdown (HasShutdownContext)
 import           Pos.Infra.Slotting.Class (MonadSlots)
 import           Pos.Infra.StateLock (StateLock, StateLockMetrics)
