+++ sinbin/cardano-sl-sinbin.cabal
@@ -17,13 +17,6 @@ cabal-version:       >=1.10
 library
   exposed-modules:
                         Pos.Sinbin.Reporting
-                        Pos.Sinbin.Reporting.Exceptions
-                        Pos.Sinbin.Reporting.Health.Types
-                        Pos.Sinbin.Reporting.Http
-                        Pos.Sinbin.Reporting.MemState
-                        Pos.Sinbin.Reporting.Methods
-                        Pos.Sinbin.Reporting.Metrics
-                        Pos.Sinbin.Reporting.Wlog
 
   other-modules:
                         Paths_cardano_sl_sinbin
@@ -39,7 +32,6 @@ library
                       , cardano-sl-core
                       , cardano-sl-crypto
                       , cardano-sl-db
-                      , cardano-sl-networking
                       , cardano-sl-util
                       , cardano-report-server >= 0.4.5
                       , conduit
