+++ sinbin/src/Pos/Sinbin/Reporting.hs
@@ -1,13 +1,5 @@
 -- | Reexports of error reporting functionality
 
 module Pos.Sinbin.Reporting
-       ( module X
+       (
        ) where
-
-import           Pos.Sinbin.Reporting.Exceptions as X
-import           Pos.Sinbin.Reporting.Health.Types as X
-import           Pos.Sinbin.Reporting.Http as X
-import           Pos.Sinbin.Reporting.MemState as X
-import           Pos.Sinbin.Reporting.Methods as X
-import           Pos.Sinbin.Reporting.Metrics as X
-import           Pos.Sinbin.Reporting.Wlog as X
