--- sinbin/src/Pos/Sinbin/Reporting/Health/Types.hs
@@ -1,9 +1,0 @@
--- | Types relevant to health reporting.
-
-module Pos.Sinbin.Reporting.Health.Types
-    ( HealthStatus (..)
-    ) where
-
-import           Data.Text (Text)
-
-data HealthStatus = HSHealthy Text | HSUnhealthy Text
