+++ ssc/src/Pos/Ssc/Logic/VAR.hs
@@ -25,6 +25,7 @@ import           Pos.Core (BlockVersionData, ComponentBlock (..),
                      HasProtocolConstants, HeaderHash, epochIndexL,
                      epochOrSlotG, headerHash)
 import           Pos.Core.Chrono (NE, NewestFirst (..), OldestFirst (..))
+import           Pos.Core.Reporting (MonadReporting, reportError)
 import           Pos.Core.Ssc (SscPayload (..))
 import           Pos.Crypto (ProtocolMagic)
 import           Pos.DB (MonadDBRead, MonadGState, SomeBatchOp (..),
@@ -33,7 +34,6 @@ import           Pos.Exception (assertionFailed)
 import           Pos.Lrc.Consumer.Ssc (getSscRichmen)
 import           Pos.Lrc.Context (HasLrcContext)
 import           Pos.Lrc.Types (RichmenStakes)
-import           Pos.Sinbin.Reporting (MonadReporting, reportError)
 import           Pos.Ssc.Configuration (HasSscConfiguration)
 import qualified Pos.Ssc.DB as DB
 import           Pos.Ssc.Error (SscVerifyError (..), sscIsCriticalVerifyError)
