+++ txp/cardano-sl-txp.cabal
@@ -66,7 +66,6 @@ library
                      , cardano-sl-crypto
                      , cardano-sl-db
                      , cardano-sl-networking
-                     , cardano-sl-sinbin
                      , cardano-sl-util
                      , conduit
                      , containers
