+++ txp/src/Pos/Txp/MemState/Class.hs
@@ -30,11 +30,11 @@ import           Data.Default (Default (..))
 import qualified Data.HashMap.Strict as HM
 import           Mockable (CurrentTime, Mockable)
 import           Pos.Core (HeaderHash)
+import           Pos.Core.Reporting (MonadReporting)
 import           Pos.Core.Slotting (MonadSlots (..))
 import           Pos.Core.Txp (TxAux, TxId)
 import           Pos.Crypto (ProtocolMagic)
 import           Pos.DB.Class (MonadDBRead, MonadGState (..))
-import           Pos.Sinbin.Reporting (MonadReporting)
 import           Pos.Txp.Configuration (HasTxpConfiguration)
 import           Pos.Txp.MemState.Types (GenericTxpLocalData (..))
 import           Pos.Txp.Toil.Failure (ToilVerFailure)
