+++ update/src/Pos/Update/Logic/Global.hs
@@ -22,13 +22,13 @@ import           Pos.Core (ApplicationName, BlockVersion, ComponentBlock (..),
                      StakeholderId, addressHash, blockVersionL, epochIndexL,
                      headerHashG, headerLeaderKeyL, headerSlotL)
 import           Pos.Core.Chrono (NE, NewestFirst, OldestFirst)
+import           Pos.Core.Reporting (MonadReporting)
 import           Pos.Core.Slotting (MonadSlotsData, SlottingData, slottingVar)
 import           Pos.Core.Update (BlockVersionData, UpId, UpdatePayload)
 import qualified Pos.DB.BatchOp as DB
 import qualified Pos.DB.Class as DB
 import           Pos.Exception (reportFatalError)
 import           Pos.Lrc.Context (HasLrcContext)
-import           Pos.Sinbin.Reporting (MonadReporting)
 import           Pos.Update.Configuration (HasUpdateConfiguration,
                      lastKnownBlockVersion)
 import           Pos.Update.DB (UpdateOp (..))
